+++ src/main/java/interpreter/math/operations/singleOperation/ExpOperation.java
@@ -15,7 +15,7 @@ public class ExpOperation extends SingleOperatorOperation {
     public MathValue evaluate() {
         int checkVal = element.checkValue();
         if(checkVal == 0){
-            double value = element.getIntValue();
+            double value = element.getDoubleValue();
             return new MathValue(exp(value));
         }
         else{
